# Conflicts:
#	src/emu/common/src/thread.cpp
#	src/emu/drivers/include/drivers/camera/camera_collection.h
#	src/emu/services/src/applist/applist.cpp
#	src/emu/services/src/sensor/sensor.cpp
#	src/emu/services/src/socket/socket.cpp
#	src/tests/epoc/CMakeLists.txt

+++ src/emu/services/src/applist/applist.cpp
@@ -674,12 +674,15 @@ namespace eka2l1 {
     }
 
     void applist_server::app_language(service::ipc_context &ctx) {
-        // AVKON asks apparc for the application language to pick which .rXX
-        // translation to load, so this must follow the configured system
-        // language -- a hardcoded English here overrides the locale for every
-        // app UI even when the ROM ships the requested translation.
+        // Apparc derives this from the phone language -- "Get application language
+        // for current phone language" in aplappinforeader.cpp, which then narrows it
+        // to the nearest localised resource file the app actually ships. Answering a
+        // constant here pins every app's UI to English whatever the locale says.
+        // EKA2L1 has no per-app resource set to narrow against, so the phone language
+        // is the whole answer; keep the old reply for the values that are not one
+        // (ELangTest and the internal "any").
         language app_lang = kern->get_current_language();
-        if ((static_cast<int>(app_lang) < static_cast<int>(language::en)) || (app_lang == language::any)) {
+        if ((app_lang < language::en) || (app_lang == language::any)) {
             app_lang = language::en;
         }
